rm dist_attr in Dtensor
