+++ paddle/phi/core/distributed/auto_parallel/reshard/reshard_function.h
@@ -17,6 +17,8 @@
 #include <vector>
 
 #include "paddle/phi/core/dense_tensor.h"
+#include "paddle/phi/core/distributed/auto_parallel/dist_attr.h"
+#include "paddle/phi/core/distributed/auto_parallel/dist_tensor.h"
 
 namespace phi {
 class DeviceContext;
